refactoring : step2 개선사항 반영
