+++ src/main/java/gift/entity/Member.java
@@ -66,4 +66,9 @@ public class Member {
         this.wishes = wishes;
     }
 
+    public boolean validatePassword(String password) {
+        return this.password
+                .equals(password);
+    }
+
 }
