+++ src/main/java/gift/exception/DatabaseAccessException.java
@@ -1,4 +1,4 @@
-package gift.exceptionhandler;
+package gift.exception;
 
 public class DatabaseAccessException extends RuntimeException {
