+++ src/main/java/gift/exception/ProductException.java
@@ -1,4 +1,4 @@
-package gift.exceptionhandler;
+package gift.exception;
 
 public class ProductException extends RuntimeException{
     public ProductException(String message) {
