+++ src/main/java/gift/exception/WishException.java
@@ -1,4 +1,4 @@
-package gift.exceptionhandler;
+package gift.exception;
 
 public class WishException extends RuntimeException{
     public WishException(String message) {
