+++ src/main/java/gift/repository/MemberRepository.java
@@ -7,5 +7,6 @@ import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByEmail(String email);
+    Optional<Member> findByEmailAndPassword(String email, String password);
 }
